requirements.txt
customtkinter
requests
psutil
wmi
Pillow